+++ library/src/methods/required/requiredAsync.test-d.ts
@@ -63,9 +63,9 @@ describe('requiredAsync', () => {
         }>();
         expectTypeOf<InferInput<Schema2>>().toEqualTypeOf<{
           key1: string;
-          key2?: number | undefined;
+          key2?: number;
           key3: string;
-          key4?: number | null | undefined;
+          key4?: number | null;
         }>();
       });
 
@@ -78,7 +78,7 @@ describe('requiredAsync', () => {
         }>();
         expectTypeOf<InferOutput<Schema2>>().toEqualTypeOf<{
           key1: string;
-          key2?: number | undefined;
+          key2?: number;
           key3: string;
           key4: number;
         }>();
@@ -131,10 +131,10 @@ describe('requiredAsync', () => {
         >();
         expectTypeOf<InferInput<Schema2>>().toEqualTypeOf<
           {
-            key1?: string | undefined;
+            key1?: string;
             key2: number;
             key3: string;
-            key4?: number | null | undefined;
+            key4?: number | null;
           } & { [key: string]: boolean }
         >();
       });
@@ -150,7 +150,7 @@ describe('requiredAsync', () => {
         >();
         expectTypeOf<InferOutput<Schema2>>().toEqualTypeOf<
           {
-            key1?: string | undefined;
+            key1?: string;
             key2: number;
             key3: string;
             key4: number;