+++ library/src/schemas/looseObject/looseObjectAsync.test-d.ts
@@ -59,8 +59,8 @@ describe('looseObjectAsync', () => {
       expectTypeOf<InferInput<Schema>>().toEqualTypeOf<
         {
           key1: string;
-          key2?: string | undefined;
-          key3?: string | null | undefined;
+          key2?: string;
+          key3?: string | null;
           key4: { key: number };
           key5: string;
         } & { [key: string]: unknown }
@@ -72,7 +72,7 @@ describe('looseObjectAsync', () => {
         {
           key1: string;
           key2: string;
-          key3?: string | null | undefined;
+          key3?: string | null;
           key4: { key: number };
           readonly key5: string;
         } & { [key: string]: unknown }