+++ library/src/schemas/objectWithRest/objectWithRestAsync.test-d.ts
@@ -67,8 +67,8 @@ describe('objectWithRestAsync', () => {
       expectTypeOf<InferInput<Schema>>().toEqualTypeOf<
         {
           key1: string;
-          key2?: string | undefined;
-          key3?: string | null | undefined;
+          key2?: string;
+          key3?: string | null;
           key4: { key: number };
           key5: string;
         } & { [key: string]: boolean }
@@ -80,7 +80,7 @@ describe('objectWithRestAsync', () => {
         {
           key1: string;
           key2: string;
-          key3?: string | null | undefined;
+          key3?: string | null;
           key4: { key: number };
           readonly key5: string;
         } & { [key: string]: boolean }