+++ library/src/types/transformation.ts
@@ -24,7 +24,7 @@ export interface BaseTransformation<
   readonly reference: (
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     ...args: any[]
-  ) => BaseTransformation<unknown, unknown, BaseIssue<unknown>>;
+  ) => BaseTransformation<TInput, TOutput, TIssue>;
   /**
    * Whether it's async.
    */
@@ -39,10 +39,10 @@ export interface BaseTransformation<
    *
    * @internal
    */
-  _run(
+  readonly _run: (
     dataset: TypedDataset<TInput, never>,
     config: Config<TIssue>
-  ): Dataset<TOutput, TIssue>;
+  ) => Dataset<TOutput, TIssue>;
   /**
    * Input, output and issue type.
    *
@@ -73,8 +73,8 @@ export interface BaseTransformationAsync<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     ...args: any[]
   ) =>
-    | BaseTransformation<unknown, unknown, BaseIssue<unknown>>
-    | BaseTransformationAsync<unknown, unknown, BaseIssue<unknown>>;
+    | BaseTransformation<TInput, TOutput, TIssue>
+    | BaseTransformationAsync<TInput, TOutput, TIssue>;
   /**
    * Whether it's async.
    */
@@ -89,17 +89,18 @@ export interface BaseTransformationAsync<
    *
    * @internal
    */
-  _run(
+  readonly _run: (
     dataset: TypedDataset<TInput, never>,
     config: Config<TIssue>
-  ): Promise<Dataset<TOutput, TIssue>>;
+  ) => Promise<Dataset<TOutput, TIssue>>;
 }
 
 /**
  * Generic transformation type.
  */
 export interface GenericTransformation<
-  TInput = unknown,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  TInput = any,
   TOutput = TInput,
   TIssue extends BaseIssue<unknown> = BaseIssue<unknown>,
 > extends BaseTransformation<TInput, TOutput, TIssue> {}
@@ -108,7 +109,8 @@ export interface GenericTransformation<
  * Generic transformation async type.
  */
 export interface GenericTransformationAsync<
-  TInput = unknown,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  TInput = any,
   TOutput = TInput,
   TIssue extends BaseIssue<unknown> = BaseIssue<unknown>,
 > extends BaseTransformationAsync<TInput, TOutput, TIssue> {}