+++ library/src/utils/_addIssue/_addIssue.ts
@@ -26,24 +26,27 @@ import { _stringify } from '../_stringify/index.ts';
 type Context =
   | BaseSchema<unknown, unknown, BaseIssue<unknown>>
   | BaseSchemaAsync<unknown, unknown, BaseIssue<unknown>>
-  | BaseValidation<unknown, unknown, BaseIssue<unknown>>
-  | BaseValidationAsync<unknown, unknown, BaseIssue<unknown>>
-  | BaseTransformation<unknown, unknown, BaseIssue<unknown>>
-  | BaseTransformationAsync<unknown, unknown, BaseIssue<unknown>>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  | BaseValidation<any, unknown, BaseIssue<unknown>>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  | BaseValidationAsync<any, unknown, BaseIssue<unknown>>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  | BaseTransformation<any, unknown, BaseIssue<unknown>>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  | BaseTransformationAsync<any, unknown, BaseIssue<unknown>>;
 
 /**
  * Other type.
  */
 interface Other<TContext extends Context> {
-  input?: unknown;
-  expected?: string;
-  received?: string;
-  message?: ErrorMessage<InferIssue<TContext>>;
-  path?: [IssuePathItem, ...IssuePathItem[]];
-  issues?: [
-    BaseIssue<InferInput<TContext>>,
-    ...BaseIssue<InferInput<TContext>>[],
-  ];
+  input?: unknown | undefined;
+  expected?: string | undefined;
+  received?: string | undefined;
+  message?: ErrorMessage<InferIssue<TContext>> | undefined;
+  path?: [IssuePathItem, ...IssuePathItem[]] | undefined;
+  issues?:
+    | [BaseIssue<InferInput<TContext>>, ...BaseIssue<InferInput<TContext>>[]]
+    | undefined;
 }
 
 /**