+++ website/src/routes/api/(types)/BaseTransformation/properties.ts
@@ -44,13 +44,17 @@ export const properties: Record<string, PropertyProps> = {
         type: 'custom',
         name: 'BaseTransformation',
         generics: [
-          'unknown',
-          'unknown',
           {
             type: 'custom',
-            name: 'BaseIssue',
-            href: '../BaseIssue/',
-            generics: ['unknown'],
+            name: 'TInput',
+          },
+          {
+            type: 'custom',
+            name: 'TOutput',
+          },
+          {
+            type: 'custom',
+            name: 'TIssue',
           },
         ],
       },